(MAINT) Implement Namespacing For Classes
This change enforces name spaces and folder and file structure more
consistantly in the module.
