+++ files/util/logger.rb
